feat(activitypub+mastoapi): add a very naive and simple search system.
as the title says, this commit adds a very simple implementation of
"/api/v2/search". it's far from complete and still is missing a lot of
functionality that is supposed to be present, but hey, it's a start.

the purpose of this commit is more to allow searching users by their
@-handles and finding AP objects based on their ID. this is already
implemented and works quite well.

should make Toki far more usable \o/